requirements.txt
pyspark==3.5.5
kafka-python==2.1.5
boto3==1.37.31